Split mod_cgi/mod_cgid into their own packages
ZC-209:  mod_cgi and mod_cgid have different use-cases, and are used with
different MPMs - cgi is for forked, and cgid is for threaded.  They also
conflict with each other.

Also, per QA testing, mod_authz_core needs to be part of the main package;
mod_authn_core probably should be the same.  Dependencies are simplified
for all the authn/authz packages.