Added "Provides: ea-apache24-exec_code_asuser" to suexec
ZC-652: With the change to ITK, I've added exec_code_asuser and/or
exec_php_asuser for each of the respective modules that perform
this action.  While nothing actually uses this, I figured it would
be nice in the future that we could create modules that don't care
how the UID changes, so long as the web server provides that
functionality.

NOTE: Having multiple modules installed each of which Provide the same
thing (e.g. suexec and suphp) doesn't break package installation.