Add in mpm dependencies in mpms directly
Case HB-1484: The EA4 code currently uses a very round about way to determine if mod_cgi
and mod_cgid should be supported or not. This change makes it so that they are REQUIRED
modules by the MPMs and conflict with eachother. This should improve the experience
of changing the MPM in the UI significantly.