Disable mod_lua by default
Case EA-4861: Added verbiage to the LoadModule config file, as well
as disabling of the module by default.  This will require the
administrator to manually enable mod_lua if they're sure they want
to use it.

This also updates the release number from 4, to 6 because we
accidentally skipped 5 when merging to master.