Add PROXY CGI patch (httpoxy) to EA4
Case ZC-2064:

This adds the httpoxy patch to EA4.  It will preserve the value
of the "Proxy:" header for most traffic, and only eliminate it
from the CGI environment variable HTTP_PROXY, a second mitigation
is offered. This patch will address this behavior in mod_cgi,
mod_cgid, mod_isapi, mod_proxy_fcgi and mod_fcgid, along with all
other consumers of httpd's built-in environment handling.