Update EA4 httpd to 2.4.23
Case EA-4872:

Brings the packaged httpd up to version 2.4.23.  As a side-effect
we had to also add --enable-proxy-fdpass to the configure script
arguemnts as the upstream configure script was fixed to properly
detect enabled / disabled options (we assumed it was enabled before)

Also this add mod_proxy_hcheck which provides a "healt check" service
for proxy daemond through httpd.  This requires mod_watchdog to be
installed, which is enforced through mod_proxy_hcheck's RPM requires.