Expose /icons/ URI on all domains
Case EA-4911:

The FancyIndexing option of autoindex.conf requires a URI to load the
associated Icons when decorating the generated index page.  Typically
this is mapped to the directory by an alias.  The choice of /icons/
masks all /icons/ requests, which has the undesirable effect of hiding
subdomain /icons/ directories (a common location for served icons).

This change rewires the FancyIndexing support to create a
/sharedApache2Icons/ alias to the shared apache2 icons directory,
and realigns all the AddIcon directives to use the new alias.