Merge ea-apache2 symlink patch and configuration control
Case EA-5487:

This enables EasyApache 4 to be reconfigured to activate the symlink
protection strategy developed by BlueHost.  This functionality should
not be enabled by default, but for those that don't have a more
effective means of detecting symlink abuse, this provides a measure of
protection at a significant performance cost.

To enable this protection, two top-level httpd configuration options
are added.  These options are not part of a standard Apache web server
configuration, so if you migrate to a different httpd offering, you will
need to remove these directives.  The added directives are:

    SymlinkProtect
    SymlinkProtectRoot

And a typical configured system with this configuration eabled
resembles

    SymlinkProtect On
    SymlinkProtectRoot /var/www/html

When SymlinkProtect is Off, SymlinkProtectRoot is not used.  Missing
directives lead to default settings equivalent to "SymlinkProtect Off"
and "SymlinkProtectRoot /var/www/html" repsectively.

Enabling this software can (and will) result in more 404 responses,
especially in the scenarios where the request eventually maps to a file
that is not owned by the user ID handling the request.