Add conflicts for prefork vs http2
As of httpd-2.4.27 Apache upstream has decided to not let http2 and
prefork work, as the only way it can work is by only allowing 1
connection per instance. This doesn't work well and can cause hanging
due to long holding on connections. Agreeing with the Apache Dev's
decision and just making them conflict as they shouldn't be used
together.