Case EA-6020: Piped logging can result in "Broken pipe" errors
When a graceful (SIGUSR1) is sent to Apache it ended up sending SIGTERM
to the piped logger. Piped loggers will die off when it sees an EOF, therefore
Apache shouldn't be killing in this manner.

Gary Stanley (gary@cpanel.net) has come up with this patch  I have
suggested he also send this upstream as this is a 10 year old bug.