Ensure that Paths configured as Aliases are exempt from SymlinkProtect checks
Case CPANEL-22257: Paths that are configured via the "Alias" directive
are now flagged as such in mod_alias. When processing requests for files
in these locations, the symlink protection checks now make an exception
and allow these files to be served successfully.

Additionally cleaned up some trailing whitespace issues in the existing
patches.