Optimize finding a module
case COBRA-10700

ap_find_linked_module was not performing
well when there a lot of IfModule statements in httpd.conf.  This change
uses a hash to keep track of modules which cut the startup time from 3s
to 1.5s on a large configuration file.  This also makes find_module
module simpler.

This is a round two of COBRA-10680 which was reverted because it didn't
match module identifiers and only handled .c files like mod_ssl.c.
This change includes matching for module identifiers like ssl_module