EA-8786: Patch apxs to use the correct value for top_builddir
The apxs that we have been shipping, when it generates a Makefile,
has been setting the "top_builddir" variable to be /etc/apache2.
This was causing the the build process to look for the *.mk files
under /etc/apache2/build when the files are actually under
/usr/lib64/apache2/build. The apxs from the httpd-devel package
supplied the CentOS repositories will set "top_builddir" to
/usr/lib64/httpd (the CentOS supplied httpd-devel places the *.mk
files under /usr/lib64/httpd/apache2). The source for the CentOS
httpd rpm includes a patch, httpd-2.4.3-apxs.patch,which changes
top_builddir to use this path. I've added a patch to our apxs
which will cause top_builddir to be set to /usr/lib64/apache2.
This causes the apxs generated module templates to build properly.