EA-10825: Fix for Ubuntu
We no longer include the 425_mod_socache_redis.conf conffile as
part of ea-apache24.  Since it is a conffile it will not be removed
on upgrade on Ubuntu.  This will cause a problem loading the
httpd.conf since it will be included and will attempt to load a
library that is no longer included with ea-apache24.  In order
for this file to be removed upon an upgrade to a version which
no longer contains it, we need to invoke
"dpkg-maintscript-helper rm_conffile" to make sure this conffile
is properly handled by Ubuntu.